feat: added badge in order and cart icon , in orders route tables , fix some bugs , but still now
