+++ src/pages/api/app/index.ts
@@ -242,6 +242,7 @@ export default async function handler(
         locations_id: {
           in: locationIds,
         },
+        is_paid: false,
       },
     });
